refactor: archive handling leaves utils.js
Zip and gzip reading, the zip writer, the CRC and the disc and ROM image
unpackers, with the file extension helper they lean on, move to archive.js.
Every importer names what it uses.

Part of #1072.
