+++ tests/unit/test-zip.js
@@ -1,9 +1,9 @@
 import { describe, it, expect } from "vitest";
 import * as fs from "fs";
-import * as utils from "../../src/utils.js";
 
 import { dirname, join } from "path";
 import { fileURLToPath } from "url";
+import { unzipDiscImage, unzipRomImage } from "../../src/archive.js";
 
 const __dirname = dirname(fileURLToPath(import.meta.url));
 
@@ -23,7 +23,7 @@ function corruptFirstEntry(zipData, offsetIntoData) {
 describe("zip tests", function () {
     it("should unzip SSD files", async () => {
         const zipData = new Uint8Array(fs.readFileSync(join(__dirname, "zip", "test-ssd.zip")));
-        const result = await utils.unzipDiscImage(zipData);
+        const result = await unzipDiscImage(zipData);
 
         expect(result.name).toBe("test.ssd");
         expect(result.data instanceof Uint8Array).toBeTruthy();
@@ -36,7 +36,7 @@ describe("zip tests", function () {
 
     it("should unzip ADFS files", async () => {
         const zipData = new Uint8Array(fs.readFileSync(join(__dirname, "zip", "test-adf.zip")));
-        const result = await utils.unzipDiscImage(zipData);
+        const result = await unzipDiscImage(zipData);
 
         expect(result.name).toBe("test.adf");
         expect(result.data instanceof Uint8Array).toBeTruthy();
@@ -49,7 +49,7 @@ describe("zip tests", function () {
 
     it("should unzip HFE files", async () => {
         const zipData = new Uint8Array(fs.readFileSync(join(__dirname, "zip", "test-hfe.zip")));
-        const result = await utils.unzipDiscImage(zipData);
+        const result = await unzipDiscImage(zipData);
 
         expect(result.name).toBe("test.hfe");
         expect(result.data instanceof Uint8Array).toBeTruthy();
@@ -62,7 +62,7 @@ describe("zip tests", function () {
 
     it("should unzip ROM files", async () => {
         const zipData = new Uint8Array(fs.readFileSync(join(__dirname, "zip", "test-rom.zip")));
-        const result = await utils.unzipRomImage(zipData);
+        const result = await unzipRomImage(zipData);
 
         expect(result.name).toBe("test.rom");
         expect(result.data instanceof Uint8Array).toBeTruthy();
@@ -75,7 +75,7 @@ describe("zip tests", function () {
 
     it("should handle ZIP with multiple files by picking the first compatible one", async () => {
         const zipData = new Uint8Array(fs.readFileSync(join(__dirname, "zip", "test-mixed.zip")));
-        const result = await utils.unzipDiscImage(zipData);
+        const result = await unzipDiscImage(zipData);
 
         // Should get the first compatible file (order may vary)
         expect(result.name === "test.ssd" || result.name === "test.rom").toBeTruthy();
@@ -83,14 +83,14 @@ describe("zip tests", function () {
     });
 
     it("should name the loadable files it passed over", async () => {
-        const result = await utils.unzipDiscImage(readZip("test-two-sides.zip"));
+        const result = await unzipDiscImage(readZip("test-two-sides.zip"));
 
         expect(result.name).toBe("side1.ssd");
         expect(result.ignored).toEqual(["side2.ssd"]);
     });
 
     it("should not count files of other kinds as passed over", async () => {
-        const result = await utils.unzipDiscImage(readZip("test-mixed.zip"));
+        const result = await unzipDiscImage(readZip("test-mixed.zip"));
 
         expect(result.name).toBe("test.ssd");
         expect(result.ignored).toEqual([]);
@@ -99,12 +99,12 @@ describe("zip tests", function () {
     it("should throw error for ZIP with no compatible files", async () => {
         const zipData = new Uint8Array(fs.readFileSync(join(__dirname, "zip", "test-ssd.zip")));
 
-        await expect(utils.unzipRomImage(zipData)).rejects.toThrow(/Couldn't find any compatible files/);
+        await expect(unzipRomImage(zipData)).rejects.toThrow(/Couldn't find any compatible files/);
     });
 
     it("should handle deflate-compressed ZIP files", async () => {
         const zipData = new Uint8Array(fs.readFileSync(join(__dirname, "zip", "test-deflated.zip")));
-        const result = await utils.unzipDiscImage(zipData);
+        const result = await unzipDiscImage(zipData);
 
         expect(result.name).toBe("test.ssd");
         expect(result.data instanceof Uint8Array).toBeTruthy();
@@ -118,36 +118,36 @@ describe("zip tests", function () {
     it("should throw for unsupported compression method", async () => {
         const zipData = new Uint8Array(fs.readFileSync(join(__dirname, "zip", "test-bzip2.zip")));
 
-        await expect(utils.unzipDiscImage(zipData)).rejects.toThrow(/Unsupported ZIP compression method/);
+        await expect(unzipDiscImage(zipData)).rejects.toThrow(/Unsupported ZIP compression method/);
     });
 
     it("should throw for truncated/corrupt ZIP files", async () => {
         const zipData = new Uint8Array(fs.readFileSync(join(__dirname, "zip", "test-truncated.zip")));
 
-        await expect(utils.unzipDiscImage(zipData)).rejects.toThrow(/Not a ZIP file/);
+        await expect(unzipDiscImage(zipData)).rejects.toThrow(/Not a ZIP file/);
     });
 
     it("should throw for non-ZIP data", async () => {
         const notZip = new Uint8Array([0, 1, 2, 3, 4, 5, 6, 7]);
 
-        await expect(utils.unzipDiscImage(notZip)).rejects.toThrow(/Not a ZIP file/);
+        await expect(unzipDiscImage(notZip)).rejects.toThrow(/Not a ZIP file/);
     });
 
     it("should throw for a stored entry whose data does not match its CRC32", async () => {
         const zipData = corruptFirstEntry(readZip("test-ssd.zip"), 0);
 
-        await expect(utils.unzipDiscImage(zipData)).rejects.toThrow(/Corrupt ZIP entry test\.ssd: CRC32 mismatch/);
+        await expect(unzipDiscImage(zipData)).rejects.toThrow(/Corrupt ZIP entry test\.ssd: CRC32 mismatch/);
     });
 
     it("should throw for a deflated entry that inflates cleanly to the wrong bytes", async () => {
         const zipData = corruptFirstEntry(readZip("test-deflated.zip"), 5);
 
-        await expect(utils.unzipDiscImage(zipData)).rejects.toThrow(/Corrupt ZIP entry test\.ssd: CRC32 mismatch/);
+        await expect(unzipDiscImage(zipData)).rejects.toThrow(/Corrupt ZIP entry test\.ssd: CRC32 mismatch/);
     });
 
     it("should throw for a deflated entry whose stream is damaged beyond inflating", async () => {
         const zipData = corruptFirstEntry(readZip("test-deflated.zip"), 0);
 
-        await expect(utils.unzipDiscImage(zipData)).rejects.toThrow(/Unable to inflate ZIP entry test\.ssd/);
+        await expect(unzipDiscImage(zipData)).rejects.toThrow(/Unable to inflate ZIP entry test\.ssd/);
     });
 });
